Add files via upload
This mask was made with Alibre Atom.  You may download a 30 day evaluation and use that. Please dont use the other more expensive versions if you do make changes and upload them so others can do the same. 

The screw used is a NATO 40mm gas mask thread.  I have not made a DIY filter yet, but this is a work in progress so if you have a gas mask filter you may try it.  

Hope this helps.

Info on material to use for filter:
https://smartairfilters.com/en/blog/best-materials-make-diy-face-mask-virus/